api/fe_interview/requirements.txt
Django>=3.2,<4
djangorestframework>=3.12,<4
django-cors-headers>=3.7,<4
